packages/newsletter-generator/pyproject.toml
[project]
name = "ai-baby-journal-newsletter-generator"
version = "0.1.0"
description = "AI を使用した保育園連絡帳生成パッケージ"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.3.0",
    "pydantic>=2.5.0",
    "jinja2>=3.1.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-python-dateutil",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/newsletter_generator"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "N", "UP", "B"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true